Backend/functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin
google-cloud-firestore
openai